+++ src/views.js/profile.js
@@ -1,6 +1,6 @@
 import { navBar } from '../components/nav-bar.js';
 import { userData } from '../components/user-data.js';
-import { editOptions } from '../components/edit-delete.js';
+// import { editOptions } from '../components/edit-delete.js';
 import { postProfile } from '../components/post-profile.js';
 
 
@@ -9,7 +9,7 @@ export const profile = () => {
   profileEl.id = 'pro';
   profileEl.appendChild(navBar());
   profileEl.appendChild(userData());
-  profileEl.appendChild(editOptions());
+  // profileEl.appendChild(editOptions());
   profileEl.appendChild(postProfile());
 
   return profileEl;
